rename terraform resources. add terraform import lines. add diff supression for deploy.content field. add import system test
